1493 - Moved function definitions outside main into utils script and updated test and test data/schema accodingly
